Refactor implementation for creating register list
